[ImageData] Fixed overflow in `scale` calculation for extreme values

+++ Source/TinyView/Models/RawImageData.cs
@@ -33,15 +33,15 @@ public class RawImageData<T> : IRawImageDataProvider where T : INumber<T>
     /// <paramref name="min"/> and <paramref name="max"/> to a 0-255 byte range.
     /// Values outside the range are clamped.
     /// </summary>
-    private void GenerateIndexedData(float min, float max)
+    private void GenerateIndexedData(double min, double max)
     {
-        float scale = min == max ? 1f : 255f / (max - min);
+        double scale = min == max ? 1.0 : 255.0 / (max - min);
         for (int y = 0; y < Height; ++y)
         {
             int offset = y * Width;
             for (int x = 0; x < Width; ++x)
             {
-                float norm = (Convert.ToSingle(_rawData[offset + x]) - min) * scale;
+                double norm = (Convert.ToDouble(_rawData[offset + x]) - min) * scale;
                 byte index = (byte)Math.Clamp(norm, 0, 255);
                 IndexedData[offset + x] = index;
             }
